requirements_web.txt
# WealthWise Pro - Flask Web App Dependencies
flask==3.1.3
bcrypt==5.0.0
mysql-connector-python==9.6.0
fpdf==1.7.2
pillow==11.1.0
